fix bug with lazyloadingcachepolicy not updating _last_updated on 304
This would happen when a request came in that would hit a not modified
content set from configcat, which would then mean every subsequent
request would also go through the same network request to configcat
to determine that config flag data had not change.

So change it that a 304 or otherwise is counted as updated
so that it reduces load and time.