move check for last updated outside of force_refresh function
As a way to keep force_refresh always updating the cache when
called, move the check for last update outside of the
force_refresh function.
This its aimed to minimize the total amount of time a number
of threads that would hit force_refresh if they were all called
at the same time.